requirements.txt
discord.py>=2.3.0
openai>=1.12.0
anthropic>=0.19.0
groq>=0.4.1
google-generativeai>=0.4.0
aiosqlite>=0.20.0
apscheduler>=3.10.4
cryptography>=42.0.5
python-dotenv>=1.0.1
aiofiles
aiohttp
tavily-python
browser-use
playwright
langchain-openai
langchain-google-genai
